+++ modules/Presentation/helpers.py
@@ -6,25 +6,28 @@ import nltk
 
 
 def save_plot(fig, name):
-    '''
-    Save Plot
+    """
+    Save Plot to file
 
     Parameters: 
+
         fig: Figure. pyplot fig. 
         name: string. Name of file. No extension. 
 
 
     Returns: 
+
         Nothing. 
-    '''
-    if (isinstance(name, str)):
+    """
+    if isinstance(name, str):
         from pathlib import Path
+
         # https://stackoverflow.com/a/273227/11255140
         # if plots folder doesn't exist, create it.
         Path("./plots").mkdir(parents=True, exist_ok=True)
         # bbox_inces="tight" ensures less white space in the image.
-        fig.savefig("./plots/"+name+".png", bbox_inces='tight')
-        print("Successfully saved Plot to ./plots/"+name+".png")
+        fig.savefig("./plots/" + name + ".png", bbox_inces="tight")
+        print("Successfully saved Plot to ./plots/" + name + ".png")
     else:
         print("Name has to be a string.")
 
@@ -34,15 +37,16 @@ def get_data_from_csv(data_path):
     Importing CSV data as pandas DataFrame.
 
     Parameters:
-    test_data_path (string): Path to the CSV data you want to import.
+
+        test_data_path (string): Path to the CSV data you want to import.
 
     Returns:
-    Pandas DataFrame
-    """
 
+        Pandas DataFrame
+    """
     test_tweets = pd.read_csv(data_path)
-    pd.set_option('display.max_columns', 500)
-    pd.set_option('display.width', 1000)
+    pd.set_option("display.max_columns", 500)
+    pd.set_option("display.width", 1000)
     return test_tweets
