+++ modules/Presentation/test_data_generator.py
@@ -19,7 +19,9 @@ def make_test_data():
     ]
     authors = ["Malte", "Camilla", "some asshole", "Asger", "Runi"]
     dates = [
-        datetime.date(2020, 5, 17) + datetime.timedelta(days=x) for x in range(100)
+        # date: year, month, day.
+        datetime.date(2020, 5, 17) + datetime.timedelta(days=x)
+        for x in range(100)
     ]  # generates range amount of dates.
     verdict = ["Positive", "Negative", "Uncertain"]
